requirements.txt
discord.py>=2.5,<3
python-dotenv>=1.0,<2
plotly>=6.1.1,<7
kaleido>=1.0,<2
certifi>=2025.10.5
pytest>=8.0,<9
